+++ docs/base/release_notes.md
@@ -1,6 +1,25 @@
 
 # Release Notes
 
+Version 0.2.0: new-interface & delete interface + tree improvements.
+-------------------------
+
+- Improvement to the whole 'show tree' function for more clarity.
+- Added following commands for creating and deleting interfaces:
+
+config
+│   ├── delete-interface
+
+new-interface
+    └── <ifname>
+        ├── cvlan-id
+        ├── ipv4address
+        ├── mtu
+        ├── netmask
+        ├── parent-interface
+        ├── status
+        └── svlan-id
+
 Version 0.1.6: bug fixes.
 -------------------------